+++ api/Controllers/StockController.cs
@@ -19,14 +19,20 @@ public class StockController : ControllerBase
     }
 
     [HttpGet]
-    public async Task<IActionResult> GetAll(){
-        var stocks = await _stockRepo.GetAllAsync();
+    public async Task<IActionResult> GetAll([FromQuery] QueryObject query){
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState);
+
+        var stocks = await _stockRepo.GetAllAsync(query);
         var stockDto = stocks.Select( s => s.ToStockDto());
         return Ok(stockDto);
     }
 
-    [HttpGet("{id}")]
+    [HttpGet("{id:int}")]
     public async Task<IActionResult> GetById([FromRoute] int id){
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState);
+        
         var stock = await _stockRepo.getByIdAsync(id);
         if(stock == null){
             return NotFound();
@@ -37,27 +43,36 @@ public class StockController : ControllerBase
     
     [HttpPost]
     public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto){
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState);
+        
         var stockModel = stockDto.ToStockFromCreateDTO();
         await _stockRepo.CreateAsync(stockModel);
         return CreatedAtAction(nameof(GetById), new {id = stockModel.Id} , stockModel.ToStockDto());
     }
 
     [HttpPut]
-    [Route("{id}")]
-    public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto updateDto ){      
-            var stockModel = await _stockRepo.UpdateAsync(id, updateDto);
+    [Route("{id:int}")]
+    public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto updateDto ){   
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState);   
+        
+        var stockModel = await _stockRepo.UpdateAsync(id, updateDto);
 
-            if (stockModel == null){
-                return NotFound();
-            }
+        if (stockModel == null){
+            return NotFound();
+        }
 
-            return Ok(stockModel.ToStockDto());
+        return Ok(stockModel.ToStockDto());
 
     }
     
     [HttpDelete]
-    [Route("{id}")]
+    [Route("{id:int}")]
     public async Task<IActionResult> Delete([FromRoute] int id){
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState);
+
         var stockModel = await _stockRepo.DeleteAsync(id);
 
         if(stockModel == null){
